ScholarBOT_anushree/requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
pypdf
reportlab
streamlit
